rag-server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
chromadb>=0.5.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.10.0
